backend/requirements.txt
# --- Web framework ---
fastapi==0.115.0
uvicorn[standard]==0.32.0
sse-starlette==2.1.3

# --- Data modeling ---
pydantic==2.9.2
pydantic-settings==2.5.2

# --- Environment ---
python-dotenv==1.0.1

# --- HTTP clients ---
requests==2.32.3
httpx==0.27.2

# --- LLM orchestration ---
langchain==0.3.3
langchain-core==0.3.10
langchain-community==0.3.2
langchain-anthropic==0.2.3
langchain-openai==0.2.2
langgraph==0.2.40

# --- Data handling ---
pandas==2.2.3
numpy==1.26.4

# --- Dev / test ---
pytest==8.3.3
pytest-asyncio==0.24.0
freezegun==1.5.1
ruff==0.6.9
